Moved some post objects from viz. to core repo.
