+++ examples/00-postprocessing/post_processing_exhaust_manifold.py
@@ -16,10 +16,9 @@ involves conjugate heat transfer.
 
 import ansys.fluent.core as pyfluent
 from ansys.fluent.core import examples
+from ansys.fluent.core.post_objects.post_objects import Graphics, Plots
 
 from ansys.fluent.visualization import set_config
-from ansys.fluent.visualization.matplotlib import Plots
-from ansys.fluent.visualization.pyvista import Graphics
 
 set_config(blocking=True, set_view_on_display="isometric")
