+++ src/ansys/fluent/visualization/matplotlib/__init__.py
@@ -1,6 +1,7 @@
 """A package that provides interfacing Fluent with Matplotlib."""
 
-from ansys.fluent.visualization.matplotlib.matplot_objects import Plots  # noqa: F401
+from ansys.fluent.core.post_objects.post_objects import Plots  # noqa: F401
+
 from ansys.fluent.visualization.matplotlib.matplot_windows_manager import (  # noqa: F401
     matplot_windows_manager,
 )
