+++ src/ansys/fluent/visualization/post_data_extractor.py
@@ -4,13 +4,15 @@ import itertools
 from typing import Dict
 
 from ansys.api.fluent.v0.field_data_pb2 import DataLocation, PayloadTag
+from ansys.fluent.core.post_objects.post_object_definitions import (
+    GraphicsDefn,
+    PlotDefn,
+)
 from ansys.fluent.core.services.field_data import (
     _FieldDataConstants,
     merge_pathlines_data,
 )
 import numpy as np
-
-from ansys.fluent.visualization.post_object_defns import GraphicsDefn, PlotDefn
 
 
 class FieldDataExtractor:
