+++ src/ansys/fluent/visualization/pyvista/__init__.py
@@ -1,6 +1,7 @@
 """A package that provides interfacing Fluent with PyVista."""
 
-from ansys.fluent.visualization.pyvista.pyvista_objects import Graphics  # noqa: F401
+from ansys.fluent.core.post_objects.post_objects import Graphics  # noqa: F401
+
 from ansys.fluent.visualization.pyvista.pyvista_windows_manager import (  # noqa: F401
     pyvista_windows_manager,
 )
